REFACTOR: Split contracts and payments for buy items
