+++ src/main/java/de/hhu/propra/sharingplatform/model/contracts/SellContract.java
@@ -1,0 +1,10 @@
+package de.hhu.propra.sharingplatform.model.contracts;
+import de.hhu.propra.sharingplatform.service.payment.IPaymentApi;
+
+public class SellContract extends Contract {
+
+    @Override
+    public void pay(IPaymentApi paymentApi){
+        String from = super.payment
+    }
+}
