+++ src/main/java/de/hhu/propra/sharingplatform/model/payments/BorrowPayment.java
@@ -1,0 +1,17 @@
+package de.hhu.propra.sharingplatform.model.payments;
+
+import de.hhu.propra.sharingplatform.model.User;
+
+public class BorrowPayment {
+
+    long amountProPayId;
+    int bail;
+    long bailProPayId;
+
+    public BorrowPayment(User sender, User recipient, int amount, int bail) {
+        this.proPayIdSender = sender.getPropayId();
+        this.proPayIdRecipient = recipient.getPropayId();
+        this.amount = amount;
+        this.bail = bail;
+    }
+}
