+++ src/main/java/de/hhu/propra/sharingplatform/service/ContractService.java
@@ -4,7 +4,7 @@ import de.hhu.propra.sharingplatform.dao.ConflictRepo;
 import de.hhu.propra.sharingplatform.dao.ContractRepo;
 import de.hhu.propra.sharingplatform.dto.Status;
 import de.hhu.propra.sharingplatform.model.Conflict;
-import de.hhu.propra.sharingplatform.model.Contract;
+import de.hhu.propra.sharingplatform.model.contracts.Contract;
 import de.hhu.propra.sharingplatform.model.Offer;
 import de.hhu.propra.sharingplatform.service.payment.IPaymentService;
 import lombok.Data;
